requirements.txt
matplotlib==3.8.3
numpy==1.26.4
Pillow==10.2.0
pytorch_lightning==2.2.1
streamlit==1.31.1
streamlit_drawable_canvas==0.9.3
torch==2.2.1
